Added proper error handling to the client
Added proper error handling for case in which the order_management
encounters an error during the processing, for example if a bad ZIP
code was provided.

In that case instead of the JSON formatted Order, it returns a message
with "status" set to "error" and a "message" that contains an error
message.

Signed-off-by: Lev Veyde <lveyde@gmail.com>